split tests changes
